+++ main.py
@@ -15,10 +15,10 @@ import src.core as core
 # to execute the tv show pipeline
 # cd <automatic-transmission-dir> & python main.py tv_show_pipeline
 #
-# statements for testing
-# media_type = "movie"
-# media_type = "tv_show"
-# media_type = "tv_season"
+#statements for testing
+#media_type = "movie"
+#media_type = "tv_show"
+#media_type = "tv_season"
 #
 # ------------------------------------------------------------------------------
 
@@ -27,10 +27,9 @@ def full_pipeline(
     debug: bool
 ):
     """
-    Full pipeline for downloading contents
-
-    :param media_type:
-    :return:
+    full pipeline for downloading contents
+    :param media_type: either "movie", "tv_show", or "tv_season
+    :param debug: param indicating whether to run in debug mode
     """
     if media_type not in ["movie", "tv_show", "tv_season"]:
         raise ValueError(f"Invalid media_type: {media_type}")
@@ -118,7 +117,6 @@ def main():
         full_pipeline(
             media_type=command_to_media[args.command],
             debug=args.debug
-
         )
     else:
         parser.print_help()